Olivia/requirements_mobile.txt
djangorestframework==3.14.0
django-cors-headers==4.3.1
